UPDATE  +state library: move data service to it's own folder within the service file. RENAME data folder in each feature and called it a service folder, UPDATE imports across the app
