+++ libs/+state/.eslintrc.json
@@ -13,7 +13,7 @@
                "error",
                {
                   "type": "attribute",
-                  "prefix": "lib",
+                  "prefix": "rya",
                   "style": "camelCase"
                }
             ],
@@ -21,7 +21,7 @@
                "error",
                {
                   "type": "element",
-                  "prefix": "lib",
+                  "prefix": "rya",
                   "style": "kebab-case"
                }
             ]
