+++ libs/+state/src/lib/facades/index.ts
@@ -1,0 +1,1 @@
+export * from './app/app.facade'
