+++ libs/+state/src/lib/services/index.ts
@@ -1,3 +1,2 @@
-export * from './data.service';
+export * from './data/data.service';
 export * from './user-location/user.service';
-export * from './view-data/view.service';
