--- libs/+state/src/lib/services/train-arrivals/index.ts
@@ -1,1 +1,0 @@
-export * from '../../adapters/train-arrivals/train-ui.adapter'
