--- libs/+state/src/lib/services/view-data/index.ts
@@ -1,1 +1,0 @@
-export * from './view.service'
