--- libs/+state/src/lib/services/view-data/view.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-   providedIn: 'root'
-})
-export class ViewService {
-   currentTabIndex = 0;
-   Loading = true;
-}
