--- libs/features/navigation/src/data/index.ts
@@ -1,1 +1,0 @@
-export * from './services/index'
