+++ libs/features/navigation/src/services/static-content/static-content.service.spec.ts
