--- libs/features/search/src/data/index.ts
@@ -1,1 +1,0 @@
-export * from './services/search.service'
