+++ libs/features/search/src/index.ts
@@ -1,3 +1,3 @@
-export * from './data'
+export * from './services'
 export * from './shared'
 export * from './ui'
