+++ libs/features/search/src/services/index.ts
@@ -1,0 +1,1 @@
+export * from './search.service'
