--- libs/features/stations/src/data/index.ts
@@ -1,3 +1,0 @@
-export * from './services/static-content/static-content.service';
-export * from './services/station-details/station-details.service';
-export * from './services/train-arrivals/train-arrivals.service';
