--- libs/features/stations/src/data/services/station-details/station-details.service.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-   providedIn: 'root'
-})
-export class StationDetailsService {
-   init(): void {
-      console.log('StationDetailsService');
-   }
-}
