+++ libs/features/stations/src/index.ts
@@ -3,7 +3,7 @@ export * from './ui/pages';
 export * from './ui/components';
 
 // Data Folder
-export * from './data/index';
+export * from './services/index';
 
 // Shared Folder
 export * from './shared';
