+++ libs/features/stations/src/services/static-content/static-content.service.ts
